refactor: 코스 만들기/수정하기 리팩토링
