+++ apps/soso-web/src/app/course/add/page.tsx
@@ -1,5 +1,17 @@
-import { redirect } from 'next/navigation'
+'use client'
+
+import { useCourseAddStore } from '@/shared/store/useCourseAddStore'
+import { useRouter } from 'next/navigation'
+import { useEffect } from 'react'
 
 export default function CourseAddPage() {
-  redirect('/course/add/shop-select')
+  const router = useRouter()
+  const reset = useCourseAddStore((s) => s.reset)
+
+  useEffect(() => {
+    reset()
+    router.replace('/course/add/shop-select')
+  }, [reset, router])
+
+  return null
 }
