+++ apps/soso-web/src/app/course/components/CourseNameEditor.tsx
