+++ apps/soso-web/src/app/course/components/ShopSearchInput.tsx
